pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "docx>=0.2.4",
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.3",
    "pyjwt>=2.10.1",
    "streamlit>=1.44.1",
    "stripe>=12.0.0",
    "uvicorn>=0.34.0",
    "weasyprint>=65.0",
]
